# internal/docs/swagger.yaml
basePath: /
definitions:
  handlers.ContentResponse:
    properties:
      content:
        type: string
    type: object
  handlers.CreatePasteRequest:
    properties:
      content:
        type: string
      expiresAt:
        type: string
    type: object
  handlers.CreateStatsRequest:
    type: object
  handlers.CreateUserRequest:
    properties:
      username:
        type: string
    type: object
  handlers.PasteCreateResponse:
    properties:
      hash:
        type: string
      id:
        type: string
      short_url:
        type: string
    type: object
  model.Paste:
    properties:
      content:
        type: string
      createdAt:
        type: string
      expiresAt:
        type: string
      hash:
        type: string
      id:
        type: string
      views:
        type: integer
    type: object
  model.ShortURL:
    properties:
      id:
        type: string
      original:
        type: string
    type: object
  model.Stats:
    properties:
      id:
        type: string
      views:
        type: integer
    type: object
  model.User:
    properties:
      id:
        type: integer
      posts:
        items:
          type: string
        type: array
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API для управления пастами, пользователями, статистикой и короткими
    URL
  title: Pastebin API
  version: "1.0"
paths:
  /api/paste:
    post:
      consumes:
      - application/json
      description: Создает новую пасту с указанным содержимым и временем истечения.
        Возвращает ID, hash и короткий URL.
      parameters:
      - description: Данные пасты
        in: body
        name: paste
        required: true
        schema:
          $ref: '#/definitions/handlers.CreatePasteRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.PasteCreateResponse'
        "400":
          description: Некорректный запрос
          schema:
            type: string
        "500":
          description: Внутренняя ошибка сервера
          schema:
            type: string
      summary: Создать новую пасту
      tags:
      - pastes
  /api/paste/{id}:
    delete:
      description: Удаляет существующую пасту по её уникальному ID
      parameters:
      - description: ID пасты
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: Паста удалена
          schema:
            type: string
        "400":
          description: Некорректный ID
          schema:
            type: string
        "404":
          description: Паста не найдена
          schema:
            type: string
      summary: Удалить пасту по ID
      tags:
      - pastes
    get:
      description: Возвращает полную информацию о пасте по её ID. Также увеличивает
        счётчик просмотров.
      parameters:
      - description: ID пасты
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Paste'
        "400":
          description: Некорректный ID
          schema:
            type: string
        "404":
          description: Паста не найдена
          schema:
            type: string
      summary: Получить пасту по ID
      tags:
      - pastes
  /api/paste/hash/{hash}:
    get:
      description: Возвращает пасту по уникальному hash. Также увеличивает счётчик
        просмотров.
      parameters:
      - description: Hash пасты
        in: path
        name: hash
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Paste'
        "400":
          description: Некорректный hash
          schema:
            type: string
        "404":
          description: Паста не найдена
          schema:
            type: string
      summary: Получить пасту по hash
      tags:
      - pastes
  /api/paste/popular:
    get:
      description: Возвращает список самых просматриваемых паст (по убыванию просмотров)
      parameters:
      - description: Максимальное количество записей (по умолчанию 5)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Paste'
            type: array
        "404":
          description: Популярные пасты не найдены
          schema:
            type: string
        "500":
          description: Ошибка при получении статистики
          schema:
            type: string
      summary: Получить популярные пасты
      tags:
      - stats
  /api/shorturl/{hash}:
    post:
      description: Создает короткую ссылку по переданному hash пасты. Используется
        первые 6 символов хэша.
      parameters:
      - description: Hash пасты
        in: path
        name: hash
        required: true
        type: string
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.ShortURL'
        "400":
          description: Хэш слишком короткий или отсутствует
          schema:
            type: string
        "500":
          description: Ошибка сервера
          schema:
            type: string
      summary: Создать короткий URL
      tags:
      - shorturls
  /api/shorturl/{id}:
    delete:
      description: Удаляет короткий URL по его ID
      parameters:
      - description: ID ShortURL
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: ShortURL удалён
          schema:
            type: string
        "400":
          description: ID отсутствует
          schema:
            type: string
        "404":
          description: ShortURL не найден
          schema:
            type: string
      summary: Удалить короткий URL
      tags:
      - shorturls
    get:
      description: Возвращает объект короткой ссылки по её ID
      parameters:
      - description: ID ShortURL
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ShortURL'
        "400":
          description: ID отсутствует
          schema:
            type: string
        "404":
          description: ShortURL не найден
          schema:
            type: string
      summary: Получить короткий URL по ID
      tags:
      - shorturls
  /api/shorturls:
    get:
      description: Возвращает список всех коротких ссылок
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.ShortURL'
            type: array
        "500":
          description: Ошибка при получении данных
          schema:
            type: string
      summary: Получить все короткие URL
      tags:
      - shorturls
  /api/stat/{id}:
    delete:
      description: Удаляет статистику просмотров по ID пасты
      parameters:
      - description: ID статистики (равен ID пасты)
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: Статистика удалена
          schema:
            type: string
        "400":
          description: ID отсутствует
          schema:
            type: string
        "404":
          description: Статистика не найдена
          schema:
            type: string
      summary: Удалить статистику
      tags:
      - stats
    get:
      description: Возвращает статистику просмотров для указанного ID пасты
      parameters:
      - description: ID статистики (равен ID пасты)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Stats'
        "400":
          description: ID отсутствует
          schema:
            type: string
        "404":
          description: Статистика не найдена
          schema:
            type: string
      summary: Получить статистику по ID
      tags:
      - stats
  /api/stats:
    get:
      description: Возвращает список всех записей статистики просмотров
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Stats'
            type: array
        "500":
          description: Ошибка сервера
          schema:
            type: string
      summary: Получить всю статистику
      tags:
      - stats
    post:
      consumes:
      - application/json
      description: Создаёт пустую запись статистики. Используется редко, т.к. обычно
        статистика создаётся автоматически.
      parameters:
      - description: Пустой объект запроса
        in: body
        name: stats
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateStatsRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Stats'
        "400":
          description: Некорректный JSON
          schema:
            type: string
        "500":
          description: Ошибка сервера
          schema:
            type: string
      summary: Создать новую запись статистики
      tags:
      - stats
  /api/user:
    post:
      consumes:
      - application/json
      description: Регистрирует нового пользователя с указанным именем
      parameters:
      - description: Тело запроса с данными пользователя
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.User'
        "400":
          description: Некорректный JSON или пустое имя
          schema:
            type: string
        "500":
          description: Ошибка сервера при создании
          schema:
            type: string
      summary: Создать нового пользователя
      tags:
      - users
  /api/user/{id}:
    delete:
      description: Удаляет пользователя по его ID
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: Пользователь успешно удалён
          schema:
            type: string
        "400":
          description: Некорректный запрос (отсутствует ID)
          schema:
            type: string
        "404":
          description: Пользователь не найден
          schema:
            type: string
      summary: Удалить пользователя
      tags:
      - users
    get:
      description: Возвращает пользователя по его уникальному ID
      parameters:
      - description: Уникальный ID пользователя
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.User'
        "400":
          description: Некорректный запрос (отсутствует ID)
          schema:
            type: string
        "404":
          description: Пользователь не найден
          schema:
            type: string
      summary: Получить пользователя по ID
      tags:
      - users
  /api/users:
    get:
      description: Возвращает список всех зарегистрированных пользователей
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.User'
            type: array
        "500":
          description: Ошибка сервера при получении пользователей
          schema:
            type: string
      summary: Получить всех пользователей
      tags:
      - users
  /s/{code}:
    get:
      description: Возвращает содержимое пасты по короткому коду (короткому URL).
        Также увеличивает счётчик просмотров.
      parameters:
      - description: Короткий код
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Контент пасты
          schema:
            $ref: '#/definitions/handlers.ContentResponse'
        "400":
          description: Код отсутствует
          schema:
            type: string
        "404":
          description: ShortURL или паста не найдена
          schema:
            type: string
      summary: Получить пасту по короткой ссылке
      tags:
      - shorturls
swagger: "2.0"